# docs/openapi.external.yaml
openapi: 3.0.3
info:
  title: Pipelogiq External API
  version: "0.3.2-preview.7"
  description: |
    External API used by SDK clients and workers.

    Pipeline creation through `/pipelines/idempotent` is the opt-in reliable
    creation contract. Its idempotency key is unique within an application and
    is retained for the lifetime of the pipeline. Reusing a key with a different
    creation request returns `409`. The legacy `/pipelines` route is preserved
    and is not idempotent.
servers:
  - url: http://localhost:8081
security:
  - ApiKeyAuth: []
paths:
  /pipelines:
    post:
      summary: Create a pipeline using the legacy non-idempotent contract
      description: |
        Preserved for existing SDK clients. Each successful request creates a
        new pipeline, even when the payload is identical. New integrations that
        must safely retry an unknown HTTP outcome should use
        `/pipelines/idempotent`.
      operationId: createPipelineLegacy
      tags:
        - Pipelines
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacyPipelineCreateRequest"
      responses:
        "200":
          description: Pipeline created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineResponse"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Invalid API key in the request body
...
  /pipelines/{pipelineId}:
    get:
      summary: Get pipeline status and stage execution details
      operationId: getPipeline
      tags:
        - Pipelines
      parameters:
        - $ref: "#/components/parameters/PipelineId"
      responses:
        "200":
          description: Pipeline status and stage details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineResponse"
        "400":
          description: Invalid pipeline ID
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Invalid API key
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Pipeline belongs to another application
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Pipeline not found
...
  /pipelines/{pipelineId}/stages:
    post:
      summary: Append stages to an existing pipeline
      operationId: appendPipelineStages
      tags:
        - Pipelines
      parameters:
        - $ref: "#/components/parameters/PipelineId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppendStagesRequest"
            examples:
              appendStages:
                value:
                  stages:
                    - stageName: agent.tool
                      stageHandlerName: AgentToolHandler
                      input: '{"toolName":"getOrder","args":{"id":"42"}}'
                      options:
                        runNextIfFailed: false
                        maxRetries: 3
                        retryInterval: 5
                        retryOnErrorCodes:
                          - TIMEOUT
                          - UPSTREAM_ERROR
                          - RATE_LIMIT_EXCEEDED
                        backoff: exponential
                        maxRetryInterval: 60
                        jitter: true
                      isEvent: false
      responses:
        "200":
          description: Stages appended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppendStagesResponse"
              examples:
                appended:
                  value:
                    stages:
                      - id: 9021
                        pipelineId: 1307
                        name: agent.tool
                        stageHandlerName: AgentToolHandler
                        status: NotStarted
                        createdAt: "2026-03-10T10:14:05Z"
                        input: '{"toolName":"getOrder","args":{"id":"42"}}'
                        nextStageId: null
                        isSkipped: false
                        isEvent: false
                        runNextIfCurrentFailed: false
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Pipeline not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Pipeline is terminal, append is not allowed
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /stages/{stageId}/lease/acquire:
    post:
      summary: Acquire the lease for a dispatched stage execution
      description: |
        Worker protocol endpoint. The worker session token binds the lease to
        the authenticated worker session. `acquired=false` is a successful
        response indicating that the delivery must not execute the handler.
      operationId: acquireStageLease
      tags:
        - Worker protocol
      security:
        - WorkerSessionAuth: []
      parameters:
        - $ref: "#/components/parameters/StageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageLeaseRequest"
      responses:
        "200":
          description: Lease decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StageLeaseResponse"
        "400":
          description: Invalid stage ID or request body
        "401":
          description: Missing or invalid worker session token
        "500":
          description: Lease operation failed
  /stages/{stageId}/lease/renew:
    post:
      summary: Renew the lease for a running stage execution
      description: |
        Worker protocol endpoint. Renewal succeeds only for the current
        execution and lease owner. The worker must stop publishing a result if
        it loses the lease.
      operationId: renewStageLease
      tags:
        - Worker protocol
      security:
        - WorkerSessionAuth: []
      parameters:
        - $ref: "#/components/parameters/StageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageLeaseRequest"
      responses:
        "200":
          description: Lease renewal decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StageLeaseResponse"
        "400":
          description: Invalid stage ID or request body
        "401":
          description: Missing or invalid worker session token
        "500":
          description: Lease operation failed
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    WorkerSessionAuth:
      type: apiKey
      in: header
      name: X-Worker-Session
      description: Session token returned by worker bootstrap.
  parameters:
    PipelineId:
      name: pipelineId
      in: path
      required: true
      description: Pipeline ID (> 0)
      schema:
        type: integer
        minimum: 1
    StageId:
      name: stageId
      in: path
      required: true
      description: Stage ID (> 0)
      schema:
        type: integer
        minimum: 1
  schemas:
    PipelineDefinition:
      type: object
      required:
        - name
        - stages
      properties:
        apiKey:
          type: string
          writeOnly: true
          deprecated: true
          description: Required only by the legacy `/pipelines` endpoint; ignored by the idempotent endpoint.
        name:
          type: string
          minLength: 1
          maxLength: 255
        traceId:
          type: string
          maxLength: 36
          description: Optional tracing identifier; not part of idempotency intent comparison.
        policies:
          type: array
          items:
            type: object
            additionalProperties: true
        stages:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/StageCreate"
        pipelineKeywords:
          type: array
          maxItems: 64
          items:
            $ref: "#/components/schemas/PipelineKeyword"
        pipelineContextItems:
          type: array
          maxItems: 64
          description: Do not use pipeline context as the consumer's sole business system of record.
          items:
            $ref: "#/components/schemas/ContextItem"
    LegacyPipelineCreateRequest:
      allOf:
        - $ref: "#/components/schemas/PipelineDefinition"
        - type: object
          required:
            - apiKey
...
    PipelineResponse:
      type: object
      required:
        - id
        - name
        - status
        - createdAt
        - isTerminal
      properties:
        id:
          type: integer
        name:
          type: string
        traceId:
          type: string
        status:
          type: string
          enum:
            - NotStarted
            - Pending
            - Running
            - Paused
            - Completed
            - Failed
            - Cancelled
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        applicationId:
          type: integer
        stageStatuses:
          type: array
          items:
            type: string
        stages:
          type: array
          items:
            $ref: "#/components/schemas/StageResponse"
        pipelineContextItems:
          type: array
          description: Sensitive values are returned as `[REDACTED]` while preserving `isSensitive=true`.
          items:
            $ref: "#/components/schemas/ContextItem"
        pipelineKeywords:
          type: array
          items:
            $ref: "#/components/schemas/PipelineKeyword"
        isEvent:
          type: boolean
        idempotencyKey:
          type: string
          description: Present for pipelines created through the idempotent contract.
        isTerminal:
          type: boolean
          description: True for Completed, Failed, or Cancelled pipeline status.
        wasExisting:
          type: boolean
          description: Compatibility field; creation outcome is authoritative in the idempotent response wrapper.
    StageCreate:
      type: object
      required:
        - stageName
        - stageHandlerName
      properties:
        stageName:
          type: string
          minLength: 1
          maxLength: 255
        stageHandlerName:
          type: string
          minLength: 1
          maxLength: 300
        description:
          type: string
          maxLength: 255
        input:
          type: string
          description: JSON-serialized input payload, at most 262144 bytes.
        policies:
          type: array
          items:
            type: object
            additionalProperties: true
        options:
          $ref: "#/components/schemas/StageOptions"
        isEvent:
          type: boolean
        runNextIfFailed:
          type: boolean
    StageInfo:
      allOf:
        - $ref: "#/components/schemas/StageCreate"
        - type: object
          properties:
            stageId:
              type: integer
              nullable: true
              description: Compatibility field from SDK, ignored by server.
            pipelineId:
              type: integer
              nullable: true
              description: Compatibility field from SDK, ignored by server.
            runNextIfCurrentFailed:
              type: boolean
              description: Compatibility alias for runNextIfFailed.
    StageOptions:
      type: object
      properties:
        runNextIfFailed:
          type: boolean
        retryInterval:
          type: integer
          minimum: 0
          description: Base retry delay in seconds.
        maxRetries:
          type: integer
          minimum: 0
          description: Maximum number of automatic retries after the first execution.
        timeOut:
          type: integer
          minimum: 1
          description: Handler timeout in seconds.
        retryOnErrorCodes:
          type: array
          description: |
            When non-empty, StageOptions retries only matching error codes.
            Terminal codes such as BUSINESS_REJECTED, VALIDATION_ERROR,
            INVALID_STATE, and MISSING_REQUIRED_DATA are never retried
            automatically.
          items:
            type: string
            minLength: 1
        backoff:
          type: string
          enum:
            - fixed
            - linear
            - exponential
          default: fixed
        maxRetryInterval:
          type: integer
          minimum: 1
          description: Maximum retry delay in seconds.
        jitter:
          type: boolean
          default: false
        dependsOn:
          type: array
          items:
            type: string
        runInParallelWith:
          type: array
          items:
            type: string
        failIfOutputEmpty:
          type: boolean
        notifyOnFailure:
          type: boolean
        runAsUser:
          type: string
    StageResponse:
      type: object
      required:
        - id
        - pipelineId
        - name
        - createdAt
        - isTerminal
      properties:
        id:
          type: integer
        pipelineId:
          type: integer
        spanId:
          type: string
        name:
          type: string
        stageHandlerName:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - NotStarted
            - Running
            - Pending
            - RetryScheduled
            - Throttled
            - WaitingForApproval
            - Completed
            - Failed
            - Skipped
            - Cancelled
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        nextRetryAt:
          type: string
          format: date-time
          nullable: true
        output:
          type: string
          nullable: true
        input:
          type: string
          nullable: true
        isSkipped:
          type: boolean
        isEvent:
          type: boolean
        nextStageId:
          type: integer
          nullable: true
        logs:
          type: array
          items:
            $ref: "#/components/schemas/StageLog"
        options:
          $ref: "#/components/schemas/StageOptions"
        failureCount:
          type: integer
          minimum: 0
        lastFailedAt:
          type: string
          format: date-time
          nullable: true
        hasFailureHistory:
          type: boolean
        attempt:
          type: integer
          minimum: 0
          description: Stable 1-based execution attempt for the current dispatched execution; zero before first dispatch.
        retryAttempt:
          type: integer
          minimum: 0
          description: Number of retries scheduled so far.
        lastErrorCode:
          type: string
          description: Last handler or timeout error code.
        failureDisposition:
          type: string
          enum:
            - retryable
            - terminal
        isTerminal:
          type: boolean
          description: True for Completed, Failed, Skipped, or Cancelled stage status.
    ContextItem:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
          minLength: 1
          maxLength: 300
        value:
          type: string
          description: At most 65536 bytes. Returned as `[REDACTED]` when isSensitive is true.
        valueType:
          type: string
        isSensitive:
          type: boolean
          default: false
          description: Marks the value for redaction from status responses and persisted logs.
    PipelineKeyword:
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
          minLength: 1
          maxLength: 300
        value:
          type: string
          minLength: 1
          maxLength: 300
    StageLog:
      type: object
      required:
        - message
        - created
      properties:
        id:
          type: integer
        stageId:
          type: integer
        message:
          type: string
        logLevel:
          type: string
        created:
          type: string
          format: date-time
    AppendStagesRequest:
      type: object
      required:
        - stages
      properties:
        stages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/StageInfo"
    AppendStagesResponse:
      type: object
      required:
        - stages
      properties:
        stages:
          type: array
          items:
            $ref: "#/components/schemas/StageDto"
    StageDto:
      type: object
      required:
        - id
        - pipelineId
        - name
        - stageHandlerName
        - status
        - createdAt
        - isSkipped
        - isEvent
        - runNextIfCurrentFailed
      properties:
        id:
          type: integer
        pipelineId:
          type: integer
        name:
          type: string
        stageHandlerName:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        input:
          type: string
        nextStageId:
          type: integer
          nullable: true
        isSkipped:
          type: boolean
        isEvent:
          type: boolean
        runNextIfCurrentFailed:
          type: boolean
        finishedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
...
    StageLeaseRequest:
      type: object
      required:
        - executionId
        - workerId
      properties:
        executionId:
          type: string
          minLength: 1
          description: Opaque execution token received in the stage delivery.
        workerId:
          type: string
          minLength: 1
          description: Worker ID associated with the authenticated worker session.
    StageLeaseResponse:
      type: object
      required:
        - acquired
      properties:
        acquired:
          type: boolean
        attempt:
          type: integer
          minimum: 1
          description: Execution attempt accepted by the server.
        leaseExpiresAt:
          type: string
          format: date-time
          nullable: true
        reason:
          type: string
          description: Machine-readable explanation when acquired is false.
    ProblemDetails:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string